NEW: ftp binary parameter download
APM vehicles support a binary parameter download
via ftp which is much faster than the conventional
parameter download. This patch adds support for
downloading and parsing the parameterfile. The
parameters are injected in the factsystem. If the
ftp download does not work, the conventional
parameter download is used as fallback.

In the majority of failures the ftp manager only stops with
"Download failed". I added more detailed results available from
the transfer, e.g. "File not found".

The APM Vehicle responds with a fixed and wrong filesize in
the MAVFTP read request response. The reason is a dynamic
file creation on the vehicle. I add an option to ignore
the filesize indicated by the vehicle for a successfull transfer.

Ramon suggested to enable the ftp parameterdownload only
for APM vehicles. The ftp download only requests parameters
for component MAV_COMP_ID_AUTOPILOT1 for APM vehicles.